Relocate: SubmissionError & ErrorType From:     org.humancellatlas.ingest.submission To:       org.humancellatlas.ingest.errors
