+++ src/main/java/org/humancellatlas/ingest/errors/ErrorType.java
@@ -1,4 +1,4 @@
-package org.humancellatlas.ingest.submission;
+package org.humancellatlas.ingest.errors;
 
 public enum ErrorType {
     WARNING,
